Move single Primitive with PointerNodes
