+++ src/main/java/org/truffleruby/extra/ffi/PointerNodes.java
@@ -56,6 +56,58 @@ public abstract class PointerNodes {
 
     }
 
+    @Primitive(name = "nativefunction_type_size", needsSelf = false, lowerFixnum = 1)
+    public static abstract class NativeFunctionTypeSizePrimitiveNode extends PrimitiveArrayArgumentsNode {
+
+        @Specialization
+        public long typeSize(int type) {
+            switch (type) {
+                case RubiniusTypes.TYPE_CHAR:
+                case RubiniusTypes.TYPE_UCHAR:
+                    return 1;
+
+                case RubiniusTypes.TYPE_SHORT:
+                case RubiniusTypes.TYPE_USHORT:
+                    return 2;
+
+                case RubiniusTypes.TYPE_INT:
+                case RubiniusTypes.TYPE_UINT:
+                    return 4;
+
+                case RubiniusTypes.TYPE_LONG:
+                case RubiniusTypes.TYPE_ULONG:
+                case RubiniusTypes.TYPE_LL:
+                case RubiniusTypes.TYPE_ULL:
+                    return 8;
+
+                case RubiniusTypes.TYPE_FLOAT:
+                    return 4;
+
+                case RubiniusTypes.TYPE_DOUBLE:
+                    return 8;
+
+                case RubiniusTypes.TYPE_PTR:
+                case RubiniusTypes.TYPE_STRPTR:
+                case RubiniusTypes.TYPE_STRING:
+                case RubiniusTypes.TYPE_CHARARR:
+                    return 8;
+
+                case RubiniusTypes.TYPE_BOOL:
+                case RubiniusTypes.TYPE_VOID:
+                case RubiniusTypes.TYPE_ENUM:
+                case RubiniusTypes.TYPE_VARARGS:
+                default:
+                    throw new UnsupportedOperationException("no type size for: " + type);
+            }
+        }
+
+        @Specialization(guards = "!isInteger(type)")
+        public Object fallback(Object type) {
+            return FAILURE;
+        }
+
+    }
+
     @Primitive(name = "pointer_set_address")
     public static abstract class PointerSetAddressPrimitiveNode extends PrimitiveArrayArgumentsNode {
